tests/requirements.txt
pytest==8.0.0
requests==2.31.0
psycopg2-binary==2.9.9
sseclient-py==1.8.0
openpyxl==3.1.2
